In QFileDialog delete a symlink rather than actual target it points to
In QFileDialog, when a resolved symlink gets deleted, the actual target
it points to gets deleted instead of the symlink itself. The patch is to
delete the symlink rather than actual target by doing the following:

1. In QFileDialog, if a directory being deleted is a resolved symlink,
do not remove the directory. Instead, call QFileSystemModel to remove
the model index.

2. In QFileSystemModel::remove(), use the full file path instead of
the resolved file path for deletion. For a symlink, delete the symlink
itself.

The patch is for Windows and Linux.

Task-number: QTBUG-29770

+++ src/widgets/dialogs/qfiledialog.cpp
@@ -3427,7 +3427,6 @@ void QFileDialogPrivate::_q_deleteCurrent()
 
         QString fileName = index.data(QFileSystemModel::FileNameRole).toString();
         QString filePath = index.data(QFileSystemModel::FilePathRole).toString();
-        bool isDir = model->isDir(index);
 
         QFile::Permissions p(index.parent().data(QFileSystemModel::FilePermissions).toInt());
 #if QT_CONFIG(messagebox)
@@ -3452,7 +3451,7 @@ void QFileDialogPrivate::_q_deleteCurrent()
             return;
 #endif // QT_CONFIG(messagebox)
 
-        if (isDir) {
+        if (model->isDir(index) && !model->fileInfo(index).isSymLink()) {
             if (!removeDirectory(filePath)) {
 #if QT_CONFIG(messagebox)
             QMessageBox::warning(q, q->windowTitle(),
